bytebuffer to 64
Former-commit-id: 3e6abba8f9b6218794769e2157fa99621108285b [formerly add18146f367d77ac88f6256c6de4f51b8c8ca5e]
Former-commit-id: 6867b7de885be7806de042fcf89251d4a696153c

+++ transactions/transactionprocessor.js
@@ -126,8 +126,8 @@ transactionprocessor.prototype.transactionFromBuffer = function (bb) {
         t.recipientId = recepient + "C";
     }
 
-    t.amount = bb.readFloat();
-    t.fee = bb.readFloat();
+    t.amount = bb.readFloat64();
+    t.fee = bb.readFloat64();
 
     var referencedTransactionBuffer = new Buffer(8);
 
@@ -177,8 +177,8 @@ transactionprocessor.prototype.transactionFromBytes = function (bytes) {
         t.recipientId = recepient + "C";
     }
 
-    t.amount = bb.readFloat();
-    t.fee = bb.readFloat();
+    t.amount = bb.readFloat64();
+    t.fee = bb.readFloat64();
 
     var referencedTransactionBuffer = new Buffer(8);
